requirements.txt
django
djangorestframework
requests
python-dotenv
pytest
psycopg2-binary
django-cors-headers
celery
redis
djangorestframework-simplejwt